TravelBlog-48: Updated laravel to 6.x.

+++ database/migrations/2019_09_20_111055_create_tags_posts_table.php
@@ -15,12 +15,12 @@ class CreateTagsPostsTable extends Migration
     public function up()
     {
         Schema::create('tags_posts', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->integer('post_id')->unsigned();
+            $table->bigInteger('post_id')->unsigned();
             $table->foreign('post_id')->references('id')->on('posts')->onDelete('cascade');
 
-            $table->integer('tag_id')->unsigned();
+            $table->bigInteger('tag_id')->unsigned();
             $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
 
             $table->string('created_at');
