+++ database/migrations/2019_09_20_135627_create_comments_table.php
@@ -15,9 +15,9 @@ class CreateCommentsTable extends Migration
     public function up()
     {
         Schema::create('comments', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
-            $table->integer('post_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('post_id')->unsigned();
             $table->integer('parent_id');
             $table->longText('body');
             $table->string('status');
@@ -25,9 +25,7 @@ class CreateCommentsTable extends Migration
             $table->string('updated_at')->nullable();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            ;
             $table->foreign('post_id')->references('id')->on('posts')->onDelete('cascade');
-            ;
         });
     }
