+++ database/migrations/2019_09_22_125645_create_stars_posts_table.php
@@ -15,12 +15,12 @@ class CreateStarsPostsTable extends Migration
     public function up()
     {
         Schema::create('stars_posts', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->integer('post_id')->unsigned();
+            $table->bigInteger('post_id')->unsigned();
             $table->foreign('post_id')->references('id')->on('posts')->onDelete('cascade');
 
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
 
             $table->string('created_at');
