+++ database/migrations/2020_04_26_082121_create_hotel_rooms_table.php
@@ -15,8 +15,8 @@ class CreateHotelRoomsTable extends Migration
     public function up()
     {
         Schema::create('hotel_rooms', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('hotel_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('hotel_id')->unsigned();
             $table->foreign('hotel_id')->references('id')->on('hotels')->onDelete('cascade');
             $table->string('price');
             $table->string('photo');
