+++ database/migrations/2020_04_26_090713_create_hotel_messages_table.php
@@ -15,8 +15,8 @@ class CreateHotelMessagesTable extends Migration
     public function up()
     {
         Schema::create('hotel_messages', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('hotel_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('hotel_id')->unsigned();
             $table->foreign('hotel_id')->references('id')->on('hotels')->onDelete('cascade');
             $table->string('name');
             $table->string('email');
